feat(svelte): FlushEditsOnHide owns the commit-on-blur safety net
One shared component force-blurs the focused element on
visibilitychange-to-hidden and pagehide, so commit-on-blur inputs write
before the page tears down (ADR-0110). Skills swaps its inline copy for
the component; Honeycrisp, which had no net, adopts it. The svelte
skill's safety-net section now points at the component instead of
prescribing per-app copies.
