+++ .agents/skills/svelte/references/mutations-and-workspace-inputs.md
@@ -212,29 +212,24 @@ The compare-then-write guard avoids a no-op Yjs transaction when focus passes th
 
 ## The safety net (app-wide, in `+layout.svelte`)
 
+Render `FlushEditsOnHide` from `@epicenter/svelte` once in the root layout
+(ADR-0110):
+
 ```svelte
 <script lang="ts">
-  function flushPendingEdits() {
-    if (
-      document.visibilityState === 'hidden' &&
-      document.activeElement instanceof HTMLElement
-    ) {
-      document.activeElement.blur();
-    }
-  }
+  import { FlushEditsOnHide } from '@epicenter/svelte';
 </script>
 
-<svelte:document onvisibilitychange={flushPendingEdits} />
-<svelte:window onpagehide={flushPendingEdits} />
+<FlushEditsOnHide />
 ```
 
-When the page is being hidden (tab close, Cmd+W, tab switch, window minimize, iOS app-switch, bfcache), `.blur()` on the focused element synchronously dispatches its blur event, which synchronously runs your commit handler, which synchronously updates the Y.Doc: all before the page tears down. Six lines, one place, every `<input onblur>` in the app inherits the resilience.
+When the page is being hidden (tab close, Cmd+W, tab switch, window minimize, iOS app-switch, bfcache), the component force-blurs the focused element; `.blur()` synchronously dispatches its blur event, which synchronously runs your commit handler, which synchronously updates the Y.Doc: all before the page tears down. One line in one place, and every `<input onblur>` in the app inherits the resilience.
 
-`visibilitychange` is a document event, `pagehide` is a window event. Per Svelte's `packages/svelte/elements.d.ts`, `onvisibilitychange` lives on `SvelteDocumentAttributes` and `onpagehide` lives on `SvelteWindowAttributes`: keep them on the right element. Listen to both: visibilitychange is more reliable on iOS Safari, pagehide catches bfcache navigations.
+Do not hand-roll the two listeners per app: `visibilitychange` is a document event and `pagehide` is a window event, and putting either on the wrong special element typechecks loosely and never fires. The component owns that split (see `packages/svelte-utils/src/flush-edits-on-hide.svelte`).
 
 ## The default for new apps
 
-Every new app under `apps/*` should ship the safety net in `+layout.svelte` as part of scaffolding. Treat it like `<Toaster />` or `<ModeWatcher />`: a layout-level concern that's free once installed. See `workspace-app-composition` for where this fits in the `+layout.svelte` checklist.
+Every app under `apps/*` ships `<FlushEditsOnHide />` in `+layout.svelte` as part of scaffolding. Treat it like `<Toaster />` or `<ModeWatcher />`: a layout-level concern that's free once installed. See `workspace-app-composition` for where this fits in the `+layout.svelte` checklist.
 
 ## When NOT to use commit-on-blur
